#160: Перенос FakeWallet/RootWallet в test/zold
